nombre de type float pris en compte

+++ headers.c
@@ -1,18 +1,18 @@
 #include <stdio.h>
 #include "headers.h"
-int add(int int1, int int2)
+float add(int int1, int int2)
 {
 	return int1 + int2;
 }
-int sous(int int1, int int2)
+float sous(int int1, int int2)
 {
 	return int1 - int2;
 }
-int diviser(int int1, int int2)
+float diviser(int int1, int int2)
 {
 	return int1 / int2;
 }
-int mult(int int1, int int2)
+float mult(int int1, int int2)
 {
 	return int1 * int2;
 }
